Annotation Accessor and Setter

These methods access and set the gene annotations stored in a AnyHermesData object.

Usage

# S4 method for class 'AnyHermesData'
annotation(object, ...)

.row_data_annotation_cols

# S4 method for class 'AnyHermesData,DataFrame'
annotation(object) <- value

Format

The annotation column names are available in the exported character vector .row_data_annotation_cols.

Arguments

object

(AnyHermesData)
object to access the annotations from.

...

not used.

value

(DataFrame)
what should the annotations be replaced with.

Value

The S4Vectors::DataFrame with the gene annotations:

symbol

desc

chromosome

size

Note

When trying to replace the required annotations with completely missing values for any genes, a warning will be given and the corresponding gene IDs will be saved in the attribute annotation.missing.genes. Note also that additional annotations beyond the required ones may be supplied and will be stored.

Examples

object <- hermes_data
head(annotation(object))
#> DataFrame with 6 rows and 4 columns
#>                        symbol                   desc  chromosome      size
#>                   <character>            <character> <character> <integer>
#> GeneID:11185             INMT indolethylamine N-me..           7      5468
#> GeneID:10677             AVIL               advillin          12     18694
#> GeneID:101928428 LOC101928428 RNA-binding protein ..  GL000220.1       138
#> GeneID:100422835      MIR3183          microRNA 3183          17        84
#> GeneID:102466731     MIR6769A         microRNA 6769a          16        73
#> GeneID:64881           PCDH20       protocadherin 20          13      5838