Control Settings for Counts Normalization

This control function allows for easy customization of the normalization settings.

control_normalize(
  log = TRUE,
  lib_sizes = NULL,
  prior_count = 1,
  fit_type = "parametric"
)

Arguments

log

(flag)
whether log2 values are returned, otherwise original scale is used.

lib_sizes

(NULL or counts)
library sizes, if NULL the vector with the sum of the counts for each of the samples will be used.

prior_count

(non-negative number)
average count to be added to each observation to avoid taking log of zero, used only when log = TRUE.

fit_type

(string)
method to estimate dispersion parameters in Negative Binomial model, used only when normalize() methods include vst and/or rlog. See estimateDispersions for details.

Value

List with the above settings used to perform the normalization procedure.

Note

To be used with the normalize() function.

Examples

control_normalize()
#> $log
#> [1] TRUE
#> 
#> $lib_sizes
#> NULL
#> 
#> $prior_count
#> [1] 1
#> 
#> $fit_type
#> [1] "parametric"
#> 
control_normalize(log = FALSE, lib_sizes = rep(1e6L, 20))
#> $log
#> [1] FALSE
#> 
#> $lib_sizes
#>  [1] 1000000 1000000 1000000 1000000 1000000 1000000 1000000 1000000 1000000
#> [10] 1000000 1000000 1000000 1000000 1000000 1000000 1000000 1000000 1000000
#> [19] 1000000 1000000
#> 
#> $prior_count
#> [1] 1
#> 
#> $fit_type
#> [1] "parametric"
#>