Extra Variable Names Accessor Methods

The methods access the names of the variables in colData() and rowData() of the object which are not required by design. So these can be additional sample or patient characteristics, or gene characteristics.

extraColDataNames(x, ...)

# S4 method for AnyHermesData
extraColDataNames(x, ...)

extraRowDataNames(x, ...)

# S4 method for AnyHermesData
extraRowDataNames(x, ...)

Arguments

x
(AnyHermesData)
object.

...
not used.

Value

The character vector with the additional variable names in either colData() or rowData().

Examples

object <- hermes_data
extraColDataNames(object)

#>  [1] "Filename" "SampleID" "AGEGRP"   "AGE18"    "STDDRS"   "STDDRSD" 
#>  [7] "STDSSDT"  "TRTDRS"   "TRTDRSD"  "SITENM"   "BHDCIRC"  "BHDCIRCU"
#> [13] "ADAFL"    "BLANP"    "BKPS"     "BLKS"     "BTANNER"  "FRPST"   
#> [19] "DURIDX"   "DURSAF"   "DURSUR"   "LNTHRPY"  "AENCIFL"  "STUDYID" 
#> [25] "USUBJID"  "SUBJID"   "RFSTDTC"  "RFENDTC"  "RFXSTDTC" "RFXENDTC"
#> [31] "RFICDTC"  "RFPENDTC" "DTHDTC"   "DTHFL"    "SITEID"   "INVID"   
#> [37] "INVNAM"   "BRTHDTC"  "AGE"      "AGEU"     "SEX"      "RACE"    
#> [43] "ETHNIC"   "ARMCD"    "ARM"      "ACTARMCD" "ACTARM"   "COUNTRY" 
#> [49] "DMDTC"    "DMDY"     "RACE1"    "RACE2"    "RACE3"    "RACE4"   
#> [55] "RACE5"    "RACE6"    "IBRTHDT"  "IBRTHDTF" "BAGE"     "BAGEU"   
#> [61] "BWT"      "BWTU"     "BHT"      "BHTU"     "BBMI"     "ITTFL"   
#> [67] "SAFFL"    "INFCODT"  "RANDDT"   "TRTSDTC"  "TRTSDTM"  "TRTSTMF" 
#> [73] "TRTEDTM"  "TRTETMF"  "TRTDUR"   "DISCSTUD" "DISCDEAT" "DISCAE"  
#> [79] "DISTRTFL" "AEWITHFL" "ALIVDT"   "COHORT"   "TTYPE"    "STDSSDY" 
#> [85] "SUBID"   

extraRowDataNames(object)

#> [1] "GeneID"          "SYMBOL"          "chromosome_name"