Predicate for Required Annotations

This helper function determines for each gene in the object whether all required annotation columns are filled.

Usage

h_has_req_annotations(object, annotation_required)

Arguments

object

(AnyHermesData)
input object.

annotation_required

(character)
names of required annotation columns for genes.

Value

Named logical vector with one value for each gene in object, which is TRUE if all required annotation columns are filled, and otherwise FALSE.

See also

Examples

object <- hermes_data
result <- h_has_req_annotations(object, "size")
all(result)
#> [1] TRUE
rowData(object)$size[1] <- NA # nolint
which(!h_has_req_annotations(object, "size"))
#> GeneID:11185 
#>            1