Inner Joining a Genes with a CDISC Data Set

This is a useful function when trying to join genetic with CDISC data sets.

inner_join_cdisc(
  gene_data,
  cdisc_data,
  patient_key = "USUBJID",
  additional_keys = character()
)

Arguments

gene_data
(data.frame or DataFrame)
genetic data.

cdisc_data
(data.frame)
CDISC data (typically patient level data).

patient_key
(string)
patient identifier.

additional_keys
(character)
potential additional keys for the two data sets.

Value

A data.frame which contains columns from both data sets merged by the keys.

Note

Columns which are contained in both data sets but are not specified as keys are taken from gene_data and not from cdisc_data.

Examples

gene_data <- col_data_with_genes(hermes_data, "counts", gene_spec("GeneID:1820"))
cdisc_data <- data.frame(
  USUBJID = head(gene_data$USUBJID, 10),
  extra = 1:10
)
result <- inner_join_cdisc(gene_data, cdisc_data)

#> Warning: Patients AB12345-CHN-7-id-28, AB12345-CHN-4-id-73, AB12345-RUS-1-id-52, AB12345-PAK-11-id-268, AB12345-CHN-13-id-102, AB12345-CHN-17-id-84, AB12345-BRA-11-id-9, AB12345-CHN-4-id-115, AB12345-CHN-15-id-245, AB12345-CHN-4-id-370 from gene data set were lost because they could not be joined to CDISC data set

result

#>                  USUBJID Filename       SampleID         AGEGRP AGE18 STDDRS
#> 1   AB12345-CHN-1-id-307     eset 06520046C0018R    >= 18 years >= 18  DEATH
#> 2  AB12345-CHN-11-id-220     eset 06520105C0017R  6 - <12 years  < 18       
#> 3  AB12345-CHN-15-id-201     eset 06520103C0017R 12 - <18 years  < 18  DEATH
#> 4  AB12345-CHN-15-id-262     eset 06520067C0018R    >= 18 years >= 18       
#> 5   AB12345-CHN-3-id-128     eset 06520011B0023R 12 - <18 years  < 18  DEATH
#> 6   AB12345-CHN-7-id-267     eset 06520092C0017R 12 - <18 years  < 18  DEATH
#> 7  AB12345-NGA-11-id-173     eset 06520062C0017R    >= 18 years >= 18  DEATH
#> 8   AB12345-RUS-3-id-378     eset 06520063C0043R  6 - <12 years  < 18  DEATH
#> 9   AB12345-USA-1-id-261     eset 06520022C0017R 12 - <18 years  < 18       
#> 10   AB12345-USA-1-id-45     eset 06520001B0023R    >= 18 years >= 18  DEATH
#>                                STDDRSD    STDSSDT              TRTDRS
#> 1  DEATH DUE TO PROGRESSION OF DISEASE 07/24/2016 PROGRESSIVE DISEASE
#> 2                                                 PROGRESSIVE DISEASE
#> 3  DEATH DUE TO PROGRESSION OF DISEASE 08/12/2016 PROGRESSIVE DISEASE
#> 4                                                 PROGRESSIVE DISEASE
#> 5  DEATH DUE TO PROGRESSION OF DISEASE 05/31/2016       ADVERSE EVENT
#> 6  DEATH DUE TO PROGRESSION OF DISEASE 02/16/2017 PROGRESSIVE DISEASE
#> 7  DEATH DUE TO PROGRESSION OF DISEASE 09/11/2016 PROGRESSIVE DISEASE
#> 8  DEATH DUE TO PROGRESSION OF DISEASE 08/05/2016 PROGRESSIVE DISEASE
#> 9                                                 PROGRESSIVE DISEASE
#> 10 DEATH DUE TO PROGRESSION OF DISEASE 01/08/2016 PROGRESSIVE DISEASE
#>                   TRTDRSD
#> 1  PROGRESSION OF DISEASE
#> 2  PROGRESSION OF DISEASE
#> 3  PROGRESSION OF DISEASE
#> 4  PROGRESSION OF DISEASE
#> 5           ADVERSE EVENT
#> 6  PROGRESSION OF DISEASE
#> 7  PROGRESSION OF DISEASE
#> 8  PROGRESSION OF DISEASE
#> 9  PROGRESSION OF DISEASE
#> 10 PROGRESSION OF DISEASE
#>                                                          SITENM BHDCIRC
#> 1   CENTRE LEON BERARD, INSTITUT D'HEMATO-ONCOLOGIE PEDIATRIQUE      NA
#> 2          UNIVERSITAETS-KINDERSPITAL; ABTEILUNG FUER ONKOLOGIE      NA
#> 3                                  DANA FARBER CANCER INSTITUTE      NA
#> 4                                  DANA FARBER CANCER INSTITUTE      NA
#> 5   CENTRE LEON BERARD, INSTITUT D'HEMATO-ONCOLOGIE PEDIATRIQUE      NA
#> 6                        MEMORIAL SLOAN KETTERING CANCER CENTER      NA
#> 7                         INSTITUT CURIE, ONCOLOGIE PEDIATRIQUE      NA
#> 8        STANFORD UNIVERSITY/LUCILE PACKARD CHILDREN'S HOSPITAL      NA
#> 9                   PENN STATE MILTON S. HERSHEY MEDICAL CENTER      NA
#> 10 FONDAZIONE IRCCS ISTITUTO NAZIONALE DEI TUMORI; STRUTTURA CO      NA
#>    BHDCIRCU ADAFL BLANP BKPS BLKS BTANNER          FRPST     DURIDX    DURSAF
#> 1               Y    NA   80   80      NA POST-MENARCHAL  61.667351 1.6755647
#> 2               Y   100   NA  100      NA  PRE-MENARCHAL  14.981520 3.0882957
#> 3               Y    NA   90   90      NA                 10.841889 0.9856263
#> 4               Y    NA  100  100      NA                131.088296 1.6755647
#> 5               Y   100   NA  100      NA                 55.030801 0.9856263
#> 6               Y   100   NA  100      NA                  7.687885 3.0225873
#> 7               Y    NA  100  100      NA                 99.285421 3.5811088
#> 8               Y    90   NA   90      NA                 93.568789 0.9856263
#> 9               Y    NA   90   90      NA                 41.626283 1.6755647
#> 10              Y    NA   90   90      NA                 45.733060 0.6570842
#>       DURSUR LNTHRPY AENCIFL STUDYID                  SUBJID          RFSTDTC
#> 1   5.158111       5      NA GO29664 GO29664-283495-11010310 2016-03-10T14:05
#> 2   5.815195       4      NA GO29664 GO29664-283694-11011201 2017-04-11T12:35
#> 3   3.055441       4      NA GO29664 GO29664-282087-11010903 2016-05-31T14:10
#> 4  19.252567       3      NA GO29664 GO29664-282087-11010102 2016-01-11T14:30
#> 5   5.026694       3      NA GO29664 GO29664-283495-11010103 2016-01-14T12:37
#> 6  12.418891       1      NA GO29664 GO29664-280959-11010901 2016-02-12T11:05
#> 7   6.340862       7      NA GO29664 GO29664-283497-11010609 2016-03-21T15:40
#> 8   4.665298       3      NA GO29664 GO29664-284024-11010804 2016-03-24T14:25
#> 9  18.858316       3      NA GO29664 GO29664-281049-11010105 2016-02-08T12:37
#> 10  2.299795       2      NA GO29664 GO29664-283971-11010601 2015-11-05T11:00
#>             RFENDTC         RFXSTDTC         RFXENDTC    RFICDTC   RFPENDTC
#> 1  2016-03-31T15:49 2016-03-10T14:05 2016-03-31T15:49 2016-02-18 2016-07-24
#> 2  2017-06-14T11:45 2017-04-11T12:35 2017-06-14T11:45 2017-04-03           
#> 3  2016-05-31T14:10 2016-05-31T14:10 2016-05-31T14:10 2016-05-11 2016-08-12
#> 4  2016-02-01T13:30 2016-01-11T14:30 2016-02-01T13:30 2016-01-07           
#> 5  2016-01-14T12:37 2016-01-14T12:37 2016-01-14T12:37 2015-12-30 2016-05-31
#> 6  2016-04-14T12:00 2016-02-12T11:05 2016-04-14T12:00 2016-02-04 2017-02-16
#> 7  2016-06-08T18:00 2016-03-21T15:40 2016-06-08T18:00 2016-03-02 2016-09-11
#> 8  2016-03-24T14:25 2016-03-24T14:25 2016-03-24T14:25 2016-03-16 2016-08-05
#> 9  2016-02-29T14:15 2016-02-08T12:37 2016-02-29T14:15 2016-02-01           
#> 10 2015-11-05T11:00 2015-11-05T11:00 2015-11-05T11:00 2015-10-30 2016-01-08
#>        DTHDTC DTHFL SITEID  INVID            INVNAM    BRTHDTC AGE  AGEU SEX
#> 1  2016-07-24     Y 283495 223804   FRAPPAZ, DIDIER 1988-05-11  27 YEARS   F
#> 2                   283694 456732   GERBER, NICOLAS 2007-01-01  10 YEARS   F
#> 3  2016-08-12     Y 282087 468105    DUBOIS, STEVEN 1999-07-13  16 YEARS   M
#> 4                   282087 468105    DUBOIS, STEVEN 1996-06-12  19 YEARS   M
#> 5  2016-05-31     Y 283495 223804   FRAPPAZ, DIDIER 2002-09-09  13 YEARS   M
#> 6  2017-02-16     Y 280959  20842   TRIPPETT, TANYA 2001-03-04  14 YEARS   M
#> 7  2016-09-11     Y 283497 241874      MICHON, JEAN 1996-12-14  19 YEARS   M
#> 8  2016-08-05     Y 284024 457432    LACAYO, NORMAN 2004-06-24  11 YEARS   M
#> 9                   281049 457179    BROWN, VALERIE 1999-07-23  16 YEARS   M
#> 10 2016-01-08     Y 283971 235545 CASANOVA, MICHELA       1996  19 YEARS   M
#>                         RACE                 ETHNIC ARMCD       ARM ACTARMCD
#> 1                    UNKNOWN           NOT REPORTED  COH3  COHORT 3     COH3
#> 2                      ASIAN     HISPANIC OR LATINO COH12 COHORT 12    COH12
#> 3                    UNKNOWN NOT HISPANIC OR LATINO COH9E COHORT 9E    COH9E
#> 4                      WHITE NOT HISPANIC OR LATINO  COH1  COHORT 1     COH1
#> 5                    UNKNOWN           NOT REPORTED  COH1  COHORT 1     COH1
#> 6  BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO COH9O COHORT 9O    COH9O
#> 7                    UNKNOWN           NOT REPORTED  COH6  COHORT 6     COH6
#> 8                    UNKNOWN     HISPANIC OR LATINO  COH8  COHORT 8     COH8
#> 9                    UNKNOWN     HISPANIC OR LATINO  COH1  COHORT 1     COH1
#> 10                     WHITE NOT HISPANIC OR LATINO  COH6  COHORT 6     COH6
#>       ACTARM COUNTRY      DMDTC DMDY RACE1 RACE2 RACE3 RACE4 RACE5 RACE6
#> 1   COHORT 3     FRA 2016-02-18  -21                NA    NA    NA    NA
#> 2  COHORT 12     SUI 2017-04-04   -7                NA    NA    NA    NA
#> 3   COHORT 9     USA 2016-05-25   -6                NA    NA    NA    NA
#> 4   COHORT 1     USA 2016-01-07   -4                NA    NA    NA    NA
#> 5   COHORT 1     FRA 2015-12-30  -15                NA    NA    NA    NA
#> 6   COHORT 9     USA 2016-02-04   -8                NA    NA    NA    NA
#> 7   COHORT 6     FRA 2016-03-02  -19                NA    NA    NA    NA
#> 8   COHORT 8     USA 2016-03-22   -2                NA    NA    NA    NA
#> 9   COHORT 1     USA 2016-02-01   -7                NA    NA    NA    NA
#> 10  COHORT 6     ITA 2015-10-30   -6                NA    NA    NA    NA
#>       IBRTHDT IBRTHDTF BAGE BAGEU    BWT BWTU   BHT BHTU     BBMI ITTFL SAFFL
#> 1  1988-05-11            27 YEARS  50.00   kg 157.0   cm 20.28480     Y     Y
#> 2  2007-01-01            10 YEARS  26.15   kg 136.0   cm 14.13819     Y     Y
#> 3  1999-07-13            16 YEARS  61.60   kg 173.0   cm 20.58204     Y     Y
#> 4  1996-06-12            19 YEARS  64.60   kg 177.1   cm 20.59659     Y     Y
#> 5  2002-09-09            13 YEARS  40.60   kg 154.0   cm 17.11924     Y     Y
#> 6  2001-03-04            14 YEARS  45.20   kg 161.1   cm 17.41596     Y     Y
#> 7  1996-12-14            19 YEARS  53.90   kg 176.0   cm 17.40057     Y     Y
#> 8  2004-06-24            11 YEARS  25.40   kg 125.0   cm 16.25600     Y     Y
#> 9  1999-07-23            16 YEARS 104.70   kg 172.0   cm 35.39075     Y     Y
#> 10 1996-06-15       MD   19 YEARS  57.00   kg 172.0   cm 19.26717     Y     Y
#>       INFCODT     RANDDT          TRTSDTC             TRTSDTM TRTSTMF
#> 1  2016-02-18 2016-03-09 2016-03-10T14:05 2016-03-10 14:05:00       S
#> 2  2017-04-03 2017-04-11 2017-04-11T12:35 2017-04-11 12:35:00       S
#> 3  2016-05-11 2016-05-31 2016-05-31T14:10 2016-05-31 14:10:00       S
#> 4  2016-01-07 2016-01-11 2016-01-11T14:30 2016-01-11 14:30:00       S
#> 5  2015-12-30 2016-01-14 2016-01-14T12:37 2016-01-14 12:37:00       S
#> 6  2016-02-04 2016-02-09 2016-02-12T11:05 2016-02-12 11:05:00       S
#> 7  2016-03-02 2016-03-21 2016-03-21T15:40 2016-03-21 15:40:00       S
#> 8  2016-03-16 2016-03-24 2016-03-24T14:25 2016-03-24 14:25:00       S
#> 9  2016-02-01 2016-02-03 2016-02-08T12:37 2016-02-08 12:37:00       S
#> 10 2015-10-30 2015-11-05 2015-11-05T11:00 2015-11-05 11:00:00       S
#>                TRTEDTM TRTETMF TRTDUR DISCSTUD DISCDEAT DISCAE DISTRTFL
#> 1  2016-03-31 16:55:59       S     22        Y        Y      N        Y
#> 2  2017-06-14 12:15:59       S     65        N        N      N        Y
#> 3  2016-05-31 15:10:59       S      1        Y        Y      N        Y
#> 4  2016-02-01 14:30:59       S     22        N        N      N        Y
#> 5  2016-01-14 13:37:59       S      1        Y        Y      N        Y
#> 6  2016-04-14 12:30:59       S     63        Y        Y      N        Y
#> 7  2016-06-08 18:30:59       S     80        Y        Y      N        Y
#> 8  2016-03-24 15:30:59       S      1        Y        Y      N        Y
#> 9  2016-02-29 14:48:59       S     22        N        N      N        Y
#> 10 2015-11-05 12:00:59       S      1        Y        Y      N        Y
#>    AEWITHFL     ALIVDT
#> 1         N 2016-07-24
#> 2         N 2017-09-27
#> 3         N 2016-08-12
#> 4         N 2017-08-15
#> 5         Y 2016-05-31
#> 6         N 2017-02-16
#> 7         N 2016-09-11
#> 8         N 2016-08-05
#> 9         N 2017-08-28
#> 10        N 2016-01-08
#>                                                           COHORT
#> 1                                       Cohort 3 (NEUROBLASTOMA)
#> 2                    Cohort 12 (ATYPICAL TERATOID RHABOID TUMOR)
#> 3  Cohort 9 (OTHER TUMOR TYPES WITH DOCUMENTED PD-L1 EXPRESSION)
#> 4                                       Cohort 1 (EWING SARCOMA)
#> 5                                       Cohort 1 (EWING SARCOMA)
#> 6  Cohort 9 (OTHER TUMOR TYPES WITH DOCUMENTED PD-L1 EXPRESSION)
#> 7                                        Cohort 6 (OSTEOSARCOMA)
#> 8                                         Cohort 8 (WILMS TUMOR)
#> 9                                       Cohort 1 (EWING SARCOMA)
#> 10                                       Cohort 6 (OSTEOSARCOMA)
#>                                                                                                                                        TTYPE
#> 1                                                                                                                              NEUROBLASTOMA
#> 2                                                                                                           ATYPICAL TERATOID RHABDOID TUMOR
#> 3                                                 GERM CELL TUMOR - YOLK SAC TUMOR (ENDODERMAL SINUS TUMOR) WITH DOCUMENTED PD-L1 EXPRESSION
#> 4                                                                                                                              EWING SARCOMA
#> 5                                                                                                                              EWING SARCOMA
#> 6  OTHER TUMOR TYPES WITH DOCUMENTED PD-L1 EXPRESSION ON EITHER TUMOR CELLS OR IMMUNE CELLS (TUMORS TYPE MUST NOT BE INCLUDED IN LIST ABOVE)
#> 7                                                                                                                               OSTEOSARCOMA
#> 8                                                                                                                                WILMS TUMOR
#> 9                                                                                                                              EWING SARCOMA
#> 10                                                                                                                              OSTEOSARCOMA
#>    STDSSDY    SUBID low_depth_flag tech_failure_flag GeneID.1820 extra
#> 1      137 11010310          FALSE             FALSE          16    10
#> 2       NA 11011201          FALSE             FALSE          54     4
#> 3       74 11010903          FALSE             FALSE         567     6
#> 4       NA 11010102          FALSE             FALSE         153     2
#> 5      139 11010103          FALSE             FALSE          49     1
#> 6      371 11010901          FALSE             FALSE         111     5
#> 7      175 11010609           TRUE             FALSE         180     9
#> 8      135 11010804           TRUE             FALSE          94     3
#> 9       NA 11010105          FALSE             FALSE         171     8
#> 10      65 11010601          FALSE             FALSE         118     7