Row Binding of AnyHermesData Objects

This method combines AnyHermesData objects with the same samples but different features of interest (rows in assays).

Arguments

...
(AnyHermesData)
objects to row bind.

Value

The combined AnyHermesData object.

Note

Note that this just inherits SummarizedExperiment::rbind,SummarizedExperiment-method(). When binding a AnyHermesData object with a SummarizedExperiment::SummarizedExperiment object, then the result will be a SummarizedExperiment::SummarizedExperiment object (the more general class).

Note that we need to have unique gene IDs (row names) and the same prefix across the combined object.

See also

Examples

a <- hermes_data[1:2542, ]
b <- hermes_data[2543:5085, ]
result <- rbind(a, b)
class(result)

#> [1] "HermesData"
#> attr(,"package")
#> [1] "hermes"