Show Method for AnyHermesData Objects

A show method that displays high-level information of AnyHermesData objects.

Usage

# S4 method for class 'HermesData'
show(object)

# S4 method for class 'RangedHermesData'
show(object)

Arguments

object

(AnyHermesData)
input.

Value

None (invisible NULL), only used for the side effect of printing to the console.

Note

The same method is used for both HermesData and RangedHermesData objects. We need to define this separately to have this method used instead of the one inherited from SummarizedExperiment::SummarizedExperiment.

Examples

object <- hermes_data
object
#> class: HermesData
#> assays(1): counts
#> genes(5085): GeneID:11185 GeneID:10677 ... GeneID:9087 GeneID:9426
#> additional gene information(3): GeneID SYMBOL chromosome_name
#> samples(20): 06520011B0023R 06520067C0018R ... 06520015C0016R
#>   06520019C0023R
#> additional sample information(72): Filename SampleID ... TTYPE STDSSDY