HermesData and RangedHermesData

The HermesData class is an extension of SummarizedExperiment::SummarizedExperiment with additional validation criteria.

HermesData(object)

HermesDataFromMatrix(counts, ...)

Arguments

object
(SummarizedExperiment)
input to create the HermesData object from. If this is a RangedSummarizedExperiment, then the result will be RangedHermesData.

counts
(matrix)
counts to create the HermesData object from.

...
additional arguments, e.g. rowData, colData, etc. passed to SummarizedExperiment::SummarizedExperiment() internally. Note that if rowRanges is passed instead of rowData, then the result will be a RangedHermesData object.

Details

The additional criteria are:

The first assay must be counts containing non-missing, integer, non-negative values.

The following columns must be in rowData:

HGNC

HGNCGeneName

Chromosome

StartBP

EndBP

WidthBP

CanonicalTranscript

ProteinTranscript

LowExpressionFlag

The following columns must be in colData:

LowDepthFlag

TechnicalFailureFlag

The object must have unique row and column names. The row names are the gene names and the column names are the sample names.

Analogously, RangedHermesData is an extension of SummarizedExperiment::RangedSummarizedExperiment and has the same additional validation requirements. Methods can be defined for both classes at the same time with the AnyHermesData signature.

A Biobase::ExpressionSet object can be imported by using the SummarizedExperiment::makeSummarizedExperimentFromExpressionSet() function to first convert it to a SummarizedExperiment::SummarizedExperiment object before converting it again into a HermesData object.

Slots

prefix

common prefix of the gene IDs (row names).

Note

Note that we use S4Vectors::setValidity2() to define the validity method, which allows us to turn off the validity checks in internal functions where intermediate objects may not be valid within the scope of the function.

It can be helpful to convert character variables to factors in colData() (before or after the HermesData creation). We provide the utility function df_char_to_factor() to simplify this task, but leave it to the user to allow for full control of the details.

Examples

# Convert an `ExpressionSet` to a `RangedSummarizedExperiment`.
ranged_summarized_experiment <- makeSummarizedExperimentFromExpressionSet(expression_set)

# Then convert to `RangedHermesData`.
HermesData(ranged_summarized_experiment)

#> class: RangedHermesData
#> assays(1): counts
#> genes(5085): GeneID:11185 GeneID:10677 ... GeneID:9087 GeneID:9426
#> additional gene information(3): GeneID SYMBOL chromosome_name
#> samples(20): 06520011B0023R 06520067C0018R ... 06520015C0016R
#>   06520019C0023R
#> additional sample information(85): Filename SampleID ... STDSSDY SUBID

# Create objects starting from a `SummarizedExperiment`.
hermes_data <- HermesData(summarized_experiment)
hermes_data

#> class: HermesData
#> assays(1): counts
#> genes(5085): GeneID:11185 GeneID:10677 ... GeneID:9087 GeneID:9426
#> additional gene information(3): GeneID SYMBOL chromosome_name
#> samples(20): 06520011B0023R 06520067C0018R ... 06520015C0016R
#>   06520019C0023R
#> additional sample information(85): Filename SampleID ... STDSSDY SUBID

# Create objects from a matrix. Note that additional arguments are not required but possible.
counts_matrix <- assay(summarized_experiment)
counts_hermes_data <- HermesDataFromMatrix(counts_matrix)