Annotation Accessor and Setter

These methods access and set the gene annotations stored in a AnyHermesData object.

# S4 method for AnyHermesData
annotation(object, ...)

.row_data_annotation_cols

# S4 method for AnyHermesData,DataFrame
annotation(object) <- value

Arguments

object
(AnyHermesData)
object to access the annotations from.

...
not used.

value
(DataFrame)
what should the annotations be replaced with.

Format

The annotation column names are available in the exported character vector .row_data_annotation_cols.

Value

The S4Vectors::DataFrame with the gene annotations:

HGNC

HGNCGeneName

Chromosome

StartBP

EndBP

WidthBP

CanonicalTranscript

ProteinTranscript

Note

When trying to replace the annotation with completely missing values for any genes, a warning will be given and the corresponding gene IDs will be saved in the attribute annotation.missing.genes.

Examples

object <- HermesData(summarized_experiment)
head(annotation(object))

#> DataFrame with 6 rows and 8 columns
#>                          HGNC                        HGNCGeneName  Chromosome
#>                   <character>                         <character> <character>
#> GeneID:11185             INMT indolethylamine N-methyltransferase           7
#> GeneID:10677             AVIL                            advillin          12
#> GeneID:101928428 LOC101928428         RNA-binding protein 25-like  GL000220.1
#> GeneID:100422835      MIR3183                       microRNA 3183          17
#> GeneID:102466731     MIR6769A                      microRNA 6769a          16
#> GeneID:64881           PCDH20                    protocadherin 20          13
#>                    StartBP     EndBP   WidthBP CanonicalTranscript
#>                  <integer> <integer> <integer>         <character>
#> GeneID:11185      30752135  30757602      5468           NM_006774
#> GeneID:10677      57797376  57816069     18694           NM_006576
#> GeneID:101928428    128808    128945       138        XM_006709932
#> GeneID:100422835   1022476   1022559        84           NR_036148
#> GeneID:102466731   4671318   4671390        73           NR_106827
#> GeneID:64881      61409685  61415522      5838           NM_022843
#>                  ProteinTranscript
#>                        <character>
#> GeneID:11185             NP_006765
#> GeneID:10677             NP_006567
#> GeneID:101928428      XP_006709995
#> GeneID:100422835                  
#> GeneID:102466731                  
#> GeneID:64881             NP_073754