DESeq2 Differential Expression Analysis

This helper functions performs the differential expression analysis with DESeq2::DESeq() for a given AnyHermesData input and design matrix.

h_diff_expr_deseq2(object, design, ...)

Arguments

object
(HermesData)
input.

design
(matrix)
design matrix.

...
additional arguments internally passed to DESeq2::DESeq() (fitType, sfType, minReplicatesForReplace, useT, minmu).

Value

A data frame with columns log2_fc (estimated log2 fold change), stat (Wald statistic), p_val (raw p-value), adj_p_pval (Benjamini-Hochberg adjusted p-value).

References

Love MI, Huber W, Anders S (2014). “Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2.” Genome Biology, 15, 550. doi: 10.1186/s13059-014-0550-8 .

Examples

object <- HermesData(summarized_experiment)

# Create the design matrix corresponding to the factor of interest.
design <- model.matrix(~SEX, colData(object))

# Then perform the `DESeq2` differential expression analysis.
result <- h_diff_expr_deseq2(object, design)
head(result)

#>               log2_fc     stat         p_val     adj_p_val
#> GeneID:8284 11.311113 23.81607 2.276737e-125 6.837042e-122
#> GeneID:8653 11.602721 21.71646 1.434198e-104 2.153449e-101
#> GeneID:8287 11.228368 21.32513 6.636919e-101  6.643556e-98
#> GeneID:7404 10.641322 20.84374  1.737070e-96  1.304105e-93
#> GeneID:7544  9.702381 18.44387  5.840928e-76  3.508061e-73
#> GeneID:6192 10.572069 15.66683  2.549966e-55  1.276258e-52

# Change of the `fitType` can be required in some cases.
result2 <- h_diff_expr_deseq2(object, design, fitType = "local")
head(result2)

#>              log2_fc     stat         p_val     adj_p_val
#> GeneID:8284 11.31288 24.45581 4.365248e-132 1.220523e-128
#> GeneID:8653 11.60317 22.29130 4.488205e-110 6.274511e-107
#> GeneID:8287 11.22922 22.13839 1.349731e-108 1.257950e-105
#> GeneID:7404 10.64261 21.72314 1.239952e-104 8.667267e-102
#> GeneID:7544  9.70351 19.13415  1.312126e-81  7.337409e-79
#> GeneID:6192 10.57244 16.20828  4.407047e-59  2.053684e-56