limma/voom Differential Expression Analysis

This helper functions performs the differential expression analysis with the voom method from the limma package (via limma::voom(), limma::lmFit() and limma::eBayes()) for given counts in a AnyHermesData object and a corresponding design matrix.

h_diff_expr_voom(object, design, ...)

Arguments

object
(AnyHermesData)
input.

design
(matrix)
design matrix.

...
additional arguments internally passed to limma::eBayes() (robust, trend, proportion, winsor.tail.p, stdev.coef.lim).

Value

A data frame with columns log2_fc (estimated log2 fold change), stat (moderated t-statistic), p_val (raw p-value), adj_p_pval (Benjamini-Hochberg adjusted p-value).

References

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). “limma powers differential expression analyses for RNA-sequencing and microarray studies.” Nucleic Acids Research, 43(7), e47. doi: 10.1093/nar/gkv007 .

Law CW, Chen Y, Shi W, Smyth GK (2014). “voom: precision weights unlock linear model analysis tools for RNA-seq read counts.” Genome Biology, 15(2), R29. doi: 10.1186/gb-2014-15-2-r29 .

Examples

object <- HermesData(summarized_experiment)

# Create the design matrix corresponding to the factor of interest.
design <- model.matrix(~SEX, colData(object))

# Then perform the differential expression analysis.
result <- h_diff_expr_voom(object, design)
head(result)

#>                 log2_fc     stat        p_val    adj_p_val
#> GeneID:8653   11.263500 25.61834 1.707755e-16 5.242762e-13
#> GeneID:8284   11.408555 25.36512 2.062050e-16 5.242762e-13
#> GeneID:9086    9.957058 22.79621 1.553519e-15 2.633215e-12
#> GeneID:246126 10.930743 22.31276 2.326134e-15 2.957098e-12
#> GeneID:7404   11.042007 21.57429 4.378847e-15 4.155683e-12
#> GeneID:8287   11.610590 21.44460 4.903460e-15 4.155683e-12

# Sometimes we might want to specify method details.
result2 <- h_diff_expr_voom(object, design, trend = TRUE, robust = TRUE)
head(result2)

#>                 log2_fc     stat        p_val    adj_p_val
#> GeneID:8653   11.263500 25.83062 9.236697e-18 2.912138e-14
#> GeneID:8284   11.408555 25.56216 1.145384e-17 2.912138e-14
#> GeneID:8287   11.610590 22.00793 2.855654e-16 3.675574e-13
#> GeneID:7404   11.042007 21.98357 2.921868e-16 3.675574e-13
#> GeneID:246126 10.930743 21.56267 3.614133e-16 3.675574e-13
#> GeneID:9086    9.957058 21.13368 5.489610e-16 4.652444e-13